requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
colorama==0.4.4
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.5
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-storages==1.12.3
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
python-dotenv==0.20.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
whitenoise==6.2.0
